database_backend/requirements.txt
django==5.2.7
psycopg2==2.9.11
djangorestframework==3.16.0